all working server tests
